feat(Laspi): Add DoAfter to Laspi internal storage
