--- Content.Server/_Starlight/Actions/Storage/PrivateStorageSystem.cs
@@ -1,7 +1,0 @@
-using Content.Shared._Starlight.Actions.Storage;
-
-namespace Content.Server._Starlight.Actions.Storage;
-
-public sealed class PrivateStorageSystem : SharedPrivateStorageSystem
-{
-}
