+++ Content.Server/_Starlight/Storage/PrivateStorageSystem.cs
@@ -1,0 +1,10 @@
+using Content.Shared._Starlight.Storage;
+
+namespace Content.Server._Starlight.Storage;
+
+/// <summary>
+/// Server-side implementation of the private storage system.
+/// </summary>
+public sealed class PrivateStorageSystem : SharedPrivateStorageSystem
+{
+}
