+++ Content.Shared/_Starlight/Storage/PrivateStorageDoAfterEvent.cs
@@ -1,7 +1,7 @@
 using Content.Shared.DoAfter;
 using Robust.Shared.Serialization;
 
-namespace Content.Shared._Starlight.Actions.Storage;
+namespace Content.Shared._Starlight.Storage;
 
 [Serializable, NetSerializable]
 public sealed partial class PrivateStorageDoAfterEvent : SimpleDoAfterEvent
